watcher/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
anthropic==0.39.0
supabase==2.9.1
psutil==6.1.0
pywin32==308; sys_platform == "win32"
pydantic==2.9.2
python-dotenv==1.0.1
